+++ src/app/(public)/resources/_components/add-new-resource-btn.tsx
@@ -198,7 +198,7 @@ export const AddNewResourceBtn = () => {
                     )}
                   </SelectValue>
                 </SelectTrigger>
-                <SelectContent>
+                <SelectContent className="max-h-[250px] overflow-y-auto">
                   {isResourceTopicsLoading ? (
                     <div className="flex items-center justify-center py-2">
                       <Loader2 className="h-4 w-4 animate-spin" />